+++ lizard_fewsnorm/models.py
@@ -429,35 +429,14 @@ class TimeSeriesCache(models.Model):
     timestepcache = models.ForeignKey(TimeStepCache)
 
     def __unicode__(self):
-        return 'loc %s, par %s, mod %s, tstep %s' % (
-            self.geolocationcache,
-            self.parametercache,
-            self.modulecache,
-            self.timestepcache)
+        return '%s,%s,%s' % (
+            self.geolocationcache.ident,
+            self.parametercache.ident,
+            self.id)
 
     def api_url(self):
         return reverse('lizard_fewsnorm_api_timeseries_detail',
                        kwargs={'id': self.id})
-
-    # def get_timeserie(self, dt_start, dt_end):
-    #     source = self.geolocationcache.fews_norm_source()
-    #     timeserieskey = source.o(self.TimeseriesKey).filter(
-    #         locationkey__id=self.geolocationcache.ident,
-    #         parameterkey__id=self.parametercache.ident,
-    #         moduleinstancekey__id=self.modulecache.ident)
-    #     if len(timeserieskey) == 0:
-    #         return []
-    #     else:
-    #         serieskey = timeserieskey[0]
-    #         return serieskey.timeseriesvaluesandflags_set.objects.filter(
-    #             datetime__gte=dt_start,
-    #             datetime__lte=dt_end)
-
-    # def get_latest_value(self, dt_start, dt_end):
-    #     timeserie = self.get_timeserie(dt_start, dt_end).order_by(
-    # '-timestep')
-    #     if len(timeserie) > 0:
-    #         return timeserie[0]
 
 
 class FewsNormSource(models.Model):